day2/persona-chat/backend/requirements.txt
fastapi
google-generativeai
python-dotenv
pydantic
fastapi-cors
